Move calculation of winner into seperate js file
